backend_sl/streetlight/requirements.txt
asgiref==3.5.2
astroid==2.15.6
asttokens==2.2.1
attrs==19.3.0
Automat==0.8.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
blinker==1.4
blis==0.2.4
Brotli==1.0.9
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cmake==3.26.3
colorama==0.4.3
comm==0.1.3
ConfigArgParse==1.5.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.0.7
cryptography==2.8
cssselect2==0.5.0
cycler==0.11.0
cymem==2.0.7
dbus-python==1.2.16
debugpy==1.6.7
decorator==4.4.2
dill==0.3.7
distro==1.4.0
Django==4.1.2
django-cors-headers==3.5.0
django-crontab==0.7.1
django-etc==1.4.0
django-globals==0.3.2
django-js-asset==2.0.0
django-mptt==0.11.0
django-thread==0.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
dlib==19.24.1
entrypoints==0.3
et-xmlfile==1.1.0
executing==1.2.0
Fiona==1.9.3
Flask==2.2.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
fonttools==4.31.2
geographiclib==2.0
geojson==3.0.1
geopandas==0.12.2
geopy==2.3.0
gevent==22.10.2
geventhttpclient==2.0.8
gmaps==0.9.0
greenlet==2.0.1
gunicorn==20.0.4
html5lib==1.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==5.0.0
importlib-resources==5.12.0
incremental==16.10.1
ipykernel==6.22.0
ipython==8.12.0
ipywidgets==8.0.6
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
keyring==18.0.1
kiwisolver==1.4.4
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.9.0
lit==16.0.5
locust==2.13.0
macaroonbakery==1.3.1
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
more-itertools==4.2.0
mpmath==1.3.0
msgpack==1.0.4
munch==2.5.0
murmurhash==1.0.9
mysql-connector==2.2.9
mysql-connector-python==8.0.30
mysqlclient==2.1.1
nest-asyncio==1.5.6
numpy==1.24.2
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.1.0
openpyxl==3.1.2
packaging==23.0
paho-mqtt==1.6.1
panda==0.3.1
pandas==2.0.0
parso==0.8.3
patsy==0.5.3
pdfkit==0.6.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.1.0
Pint==0.20.1
plac==0.9.6
platformdirs==3.2.0
plotly==5.14.1
plotly-express==0.4.1
preshed==3.0.8
prompt-toolkit==3.0.38
protobuf==3.20.1
psutil==5.9.1
pure-eval==0.2.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
pydantic==1.10.7
pydyf==0.1.2
pygame==2.3.0
Pygments==2.15.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==2.3.0
pylint==2.17.5
pymacaroons==0.13.0
pymongo==3.13.0
PyMySQL==0.10.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyphen==0.12.0
pyproj==3.5.0
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
pytz==2020.1
PyVirtualDisplay==3.0
PyYAML==5.3.1
pyzmq==24.0.1
razorpay==1.3.0
requests==2.28.1
requests-unixsocket==0.2.0
roundrobin==0.0.4
schedule==1.1.0
scipy==1.10.1
SecretStorage==2.3.1
service-identity==18.1.0
shapely==2.0.1
simplejson==3.16.0
six==1.14.0
smart-open==6.3.0
soupsieve==2.5
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.2.19
sqlparse==0.2.4
srsly==1.0.6
ssh-import-id==5.10
stack-data==0.6.2
statsmodels==0.13.5
sympy==1.12
systemd-python==234
tenacity==8.2.2
thread6==0.2.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tomli==2.0.1
tomlkit==0.12.1
tornado==6.3
tqdm==4.65.0
traitlets==5.9.0
twilio==7.8.2
Twisted==18.9.0
typer==0.7.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.25.8
wadllib==1.3.3
wasabi==0.10.1
wcwidth==0.2.6
weasyprint==54.3
webencodings==0.5.1
whitenoise==5.2.0
widgetsnbextension==4.0.7
wrapt==1.15.0
zipp==3.15.0
zope.event==4.5.0
zopfli==0.2.1